separated: writer into its own module and added a helper function that writes to file
