+++ crates/lib/disolv-output/src/result.rs
@@ -1,13 +1,7 @@
 use std::fs;
-use std::fs::File;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-use arrow::csv::Writer;
-use arrow::datatypes::{Schema, SchemaRef};
-use arrow::record_batch::RecordBatchWriter;
-use parquet::arrow::ArrowWriter;
-use parquet::basic::Compression;
-use parquet::file::properties::WriterProperties;
+use arrow::datatypes::Schema;
 use serde::Deserialize;
 
 use disolv_core::bucket::TimeMS;
@@ -50,74 +44,6 @@ pub trait ResultWriter {
     fn write_to_file(&mut self);
     fn close_file(self);
 }
-
-#[derive(Debug)]
-pub enum WriterType {
-    Parquet(WriterParquet),
-    Csv(WriterCsv),
-}
-
-impl WriterType {
-    pub fn new(file_name: &PathBuf, schema: Schema) -> Self {
-        if file_name.exists() {
-            match std::fs::remove_file(file_name) {
-                Ok(_) => {}
-                Err(e) => panic!("Error deleting file: {}", e),
-            }
-        }
-        match file_name.extension() {
-            Some(ext) => match ext.to_str() {
-                Some("parquet") => WriterType::Parquet(WriterParquet::new(file_name, schema)),
-                Some("csv") => WriterType::Csv(WriterCsv::new(file_name)),
-                _ => panic!("Invalid file extension"),
-            },
-            None => panic!("Invalid file extension"),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct WriterParquet {
-    pub writer: ArrowWriter<File>,
-}
-
-impl WriterParquet {
-    fn new(file_name: &PathBuf, schema: Schema) -> Self {
-        let props = WriterProperties::builder()
-            .set_compression(Compression::SNAPPY)
-            .build();
-        let output_file = match File::create(file_name) {
-            Ok(file) => file,
-            Err(_) => panic!("Failed to create links file to write"),
-        };
-        let writer = match ArrowWriter::try_new(output_file, SchemaRef::from(schema), Some(props)) {
-            Ok(writer) => writer,
-            Err(_) => panic!("Failed to create links file writer"),
-        };
-        Self { writer }
-    }
-
-    pub fn close(self) {
-        self.writer.close().expect("Failed to close parquet file");
-    }
-}
-
-#[derive(Debug)]
-pub struct WriterCsv {
-    pub writer: Writer<File>,
-}
-
-impl WriterCsv {
-    fn new(file_name: &PathBuf) -> Self {
-        let writer = Writer::new(File::create(file_name).expect("failed to create file"));
-        Self { writer }
-    }
-
-    pub fn close(self) {
-        self.writer.close().expect("failed to close csv file");
-    }
-}
-
 #[derive(Debug)]
 pub struct Results {
     pub positions: Option<PositionWriter>,
