+++ crates/lib/disolv-output/src/tables/model.rs
@@ -6,7 +6,8 @@ use arrow::array::{ArrayRef, Float32Array, RecordBatch, StringArray, UInt64Array
 use arrow::datatypes::{DataType, Field, Schema};
 use typed_builder::TypedBuilder;
 
-use crate::result::{ResultWriter, WriterType};
+use crate::result::ResultWriter;
+use crate::writer::WriterType;
 
 #[derive(Debug, TypedBuilder)]
 pub struct ModelUpdate {
@@ -111,20 +112,7 @@ impl ResultWriter for ModelTrace {
             ),
         ])
         .expect("Failed to convert results to record batch");
-        match &mut self.to_output {
-            WriterType::Parquet(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write parquet");
-            }
-            WriterType::Csv(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write csv");
-            }
-        }
+        self.to_output.record_batch_to_file(&record_batch);
     }
 
     fn close_file(self) {
