+++ crates/lib/disolv-output/src/tables/payload.rs
@@ -9,7 +9,8 @@ use typed_builder::TypedBuilder;
 use disolv_core::agent::AgentId;
 use disolv_core::bucket::TimeMS;
 
-use crate::result::{ResultWriter, WriterType};
+use crate::result::ResultWriter;
+use crate::writer::WriterType;
 
 #[derive(Debug, TypedBuilder)]
 pub struct PayloadUpdate {
@@ -105,20 +106,7 @@ impl ResultWriter for PayloadTraceWriter {
             ),
         ])
         .expect("Failed to convert results to record batch");
-        match &mut self.to_output {
-            WriterType::Parquet(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write parquet");
-            }
-            WriterType::Csv(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write csv");
-            }
-        }
+        self.to_output.record_batch_to_file(&record_batch);
     }
 
     fn close_file(self) {
