+++ crates/lib/disolv-output/src/tables/tx.rs
@@ -7,7 +7,8 @@ use typed_builder::TypedBuilder;
 
 use disolv_core::bucket::TimeMS;
 
-use crate::result::{ResultWriter, WriterType};
+use crate::result::ResultWriter;
+use crate::writer::WriterType;
 
 #[derive(TypedBuilder, Copy, Clone)]
 pub struct TxData {
@@ -147,20 +148,7 @@ impl ResultWriter for TxDataWriter {
             ),
         ])
         .expect("Failed to convert results to record batch");
-        match &mut self.to_output {
-            WriterType::Parquet(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write parquet");
-            }
-            WriterType::Csv(to_output) => {
-                to_output
-                    .writer
-                    .write(&record_batch)
-                    .expect("Failed to write csv");
-            }
-        }
+        self.to_output.record_batch_to_file(&record_batch);
     }
 
     fn close_file(self) {
